refactor(terminal): move terminal + manager to src/components/terminal/
These modules are used by both project mode and home view, so they
don't belong under project/. New structure: terminal.ts, manager.ts,
and barrel index.ts under src/components/terminal/.
