+++ src/components/homeView.ts
@@ -8,8 +8,8 @@ import type { Project, PtySpawnOptions } from '../types';
 import { homeViewActive, projectPath } from './project/signals';
 import { ensureHiddenSessionsContainer } from './project/state';
 import { exitProjectMode } from './project/projectMode';
-import { OuijitTerminal, scrollSafeFit, resolveTerminalLabel } from './project/terminal';
-import { getManager } from './project/terminalManager';
+import { OuijitTerminal, scrollSafeFit, resolveTerminalLabel } from './terminal';
+import { getManager } from './terminal';
 import { setupCardActions, reconnectTerminal, collapseTagInput } from './project/terminalCards';
 import { convertIconsIn } from '../utils/icons';
 import { escapeHtml } from '../utils/html';
