+++ src/components/project/terminalCards.ts
@@ -17,8 +17,8 @@ import type { SummaryType } from './state';
 import { STACK_PAGE_SIZE, projectState } from './state';
 import { hideRunnerPanel, projectRegistry } from './helpers';
 import { projectPath, projectData, invalidateTaskList, homeViewActive } from './signals';
-import { OuijitTerminal, scrollSafeFit, resolveTerminalLabel } from './terminal';
-import { getManager } from './terminalManager';
+import { OuijitTerminal, scrollSafeFit, resolveTerminalLabel } from '../terminal';
+import { getManager } from '../terminal';
 import { showToast } from '../importDialog';
 import { showHookConfigDialog } from '../hookConfigDialog';
 import { refreshTerminalGitStatus } from './gitStatus';
