+++ src/components/terminal/index.ts
@@ -1,0 +1,4 @@
+export { OuijitTerminal, scrollSafeFit, resolveTerminalLabel, formatBranchNameForDisplay } from './terminal';
+export type { TerminalOptions } from './terminal';
+export { getManager, TerminalManager } from './manager';
+export type { StoredSession } from './manager';
